+++ CHANGES.rst
@@ -37,6 +37,7 @@
 * ``render_with`` no longer offers a shorthand for JSONP responses
 * ``coaster.sqlalchemy.ModelBase`` now replaces Flask-SQLAlchemy's db.Model
   with full support for type hinting
+* New: ``coaster.assets.WebpackManifest`` provides Webpack assets in Jinja2
 
 0.6.1 - 2021-01-06
 ------------------